requirements.txt
annotated-types==0.6.0
anyio==4.2.0
CacheControl==0.14.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
dnspython==2.5.0
email-validator==2.1.0.post1
fastapi==0.109.2
firebase-admin==6.4.0
google-api-core==2.16.2
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
msgpack==1.0.7
orjson==3.9.13
packaging==23.2
pluggy==1.4.0
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.0
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
pytest==8.0.0
python-dotenv==1.0.1
python-multipart==0.0.7
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
sentry-sdk==1.40.0
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
ujson==5.9.0
uritemplate==4.1.1
urllib3==2.2.0
uvicorn==0.27.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
